+++ packages/adhere-ui-datedisplay/es/datedisplay.d.ts
@@ -1,0 +1,5 @@
+import 'dayjs/locale/zh-cn';
+import 'dayjs/locale/pt';
+import 'dayjs/locale/en';
+declare const Components: {};
+export default Components;