+++ packages/adhere-ui-datedisplay/es/index.d.ts
@@ -1,0 +1,2 @@
+import DateDisplay from './datedisplay';
+export default DateDisplay;