+++ packages/adhere-ui-datedisplay/src/types.ts
@@ -1,0 +1,10 @@
+import React from 'react';
+
+/**
+ * IDateDisplayProps
+ * @interface IDateDisplayProps
+ */
+export interface IDateDisplayProps {
+  className?: string;
+  style?: React.CSSProperties;
+}