+++ packages/adhere/script/build.js
@@ -61,6 +61,7 @@ const namedMap = new Map([
   ['@baifendian/adhere-ui-tablegridlayout', 'TableGridLayout'],
   ['@baifendian/adhere-ui-writingboard', 'WritingBoard'],
   ['@baifendian/adhere-ui-contourblock', 'ContourBlock'],
+  ['@baifendian/adhere-ui-datedisplay', 'DateDisplay'],
   // ['@baifendian/adhere-ui-bmap', 'BMap'],
   ['@baifendian/adhere-ui-forceupdate', 'ForceUpdate'],
   ['@baifendian/adhere-util', 'Util'],