+++ packages/adhere/src/index.ts
@@ -42,6 +42,7 @@ import TableHeadSearch from '@baifendian/adhere-ui-tableheadsearch';
 import TableList from '@baifendian/adhere-ui-tablelist';
 import WritingBoard from '@baifendian/adhere-ui-writingboard';
 import ContourBlock from '@baifendian/adhere-ui-contourblock';
+import DateDisplay from '@baifendian/adhere-ui-datedisplay';
 import Util from '@baifendian/adhere-util';
 import AdapterScreen from '@baifendian/adhere-util-adapterscreen';
 import Browsersniff from '@baifendian/adhere-util-browsersniff';
@@ -103,6 +104,7 @@ export {
   TableList,
   WritingBoard,
   ContourBlock,
+  DateDisplay,
   Util,
   AdapterScreen,
   Browsersniff,